requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-decouple
apscheduler
requests
pydantic
EOF